Do not wake up unnecessarily
To simultaneously check for window events from Sway and config updates,
we were polling (check both, then sleep 100ms, in a loop). This means we
always wake up ten times per second, which is inefficient: although the
CPU usage should not be too noticeable on a modern CPU, it's not great
for power efficiency on an otherwise idle system.

Improve matters by asynchronously checking for config updates:

- Switch from the async-std and futures crates to async-io and
  futures-lite. These are what swayipc-async uses under the hood,
  and they provide all functionality we need: we might as well use them.
- Use async-io to poll inotify asynchronously, producing a stream of
  config updates.
- Merge this with the stream of events we already have, and process both
  in parallel.

This results in sworkstyle being completely idle between window
events (and it should respond slightly faster to them as it no longer
needs to wait an average 50ms in the 100ms sleep between events).