Refactor ConfigSource to require Inotify
We have no current or planned use for the config path if it did not
exist at startup, so there is no point in keeping it around.

Although it's not much shorter, this removes two levels of indentation
from the `stream::try_unfold` call (by removing an `if` and simplifying
the signature of the closure).